slightly improve the NNUE SIMD speed
--------------------------------------------------
Results of dev vs base (1+0.01, 1t, 16MB, UHO_Lichess_4852_v1.epd):
Elo: 4.92 +/- 2.96, nElo: 9.13 +/- 5.49
LOS: 99.94 %, DrawRatio: 51.21 %, PairsRatio: 1.11
Games: 15402, Wins: 4025, Losses: 3807, Draws: 7570, Points: 7810.0 (50.71 %)
Ptnml(0-2): [105, 1679, 3944, 1839, 134], WL/DD Ratio: 0.95
LLR: 2.92 (101.0%) (-2.25, 2.89) [0.00, 3.00]
--------------------------------------------------

+++ lib/nnue.rs
@@ -28,10 +28,10 @@ pub use value::*;
 pub const FTQ: i16 = 255;
 
 /// Quantization constant for the hidden layers.
-pub const HLQ: i32 = 127;
+pub const HLQ: i16 = 127;
 
 /// Quantization scale for the hidden layers.
-pub const HLS: i32 = 64;
+pub const HLS: i16 = 64;
 
 const unsafe fn copy_bytes<T>(dst: &mut T, src: &[u8]) -> usize {
     let len = size_of_val(dst);
